# api/swagger/swagger.yaml
swagger: '2.0'
info:
  description: Communication with Slack API
  version: '1'
  title: Slack API
host: localhost:10010
basePath: /
schemes:
  - https
paths:
...
  /sendMsg:
    x-swagger-router-controller: sendMsg
    post:
      operationId: sendMsg
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              text:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /sendMsg/oneButton:
    x-swagger-router-controller: sendMsg
    post:
      operationId: sendMsgOneButton
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              text:
                type: string
              callbackId:
                type: string 
              textButton1:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /sendMsg/oneButton/confirm:
    x-swagger-router-controller: sendMsg
    post:
      operationId: sendMsgOneButtonConfirm
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              text:
                type: string
              callbackId:
                type: string 
              textButton1:
                type: string
              textConfirmation1:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /sendMsg/twoButtons:
    x-swagger-router-controller: sendMsg
    post:
      operationId: sendMsgTwoButtons
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              text:
                type: string
              callbackId:
                type: string 
              textButton1:
                type: string
              textButton2:
                type: string

      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /sendMsg/twoButtons/Confirm:
    x-swagger-router-controller: sendMsg
    post:
      operationId: sendMsgTwoButtonsConfirm
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              text:
                type: string
              callbackId:
                type: string 
              textButton1:
                type: string
              textButton2:
                type: string
              textConfirmation1:
                type: string
              textConfirmation2:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /sendMsg/ephemeral:
    x-swagger-router-controller: sendMsgEphemeral
    post:
      operationId: sendMsgEphemeral
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              user:
                type: string
              text:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /sendMsg/ephemeral/oneButton/:
    x-swagger-router-controller: sendMsgEphemeral
    post:
      operationId: sendMsgEphemeralOneButton
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              user:
                type: string
              text:
                type: string
              callbackId:
                type: string 
              textButton1:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /sendMsg/ephemeral/oneButton/confirm:
    x-swagger-router-controller: sendMsgEphemeral
    post:
      operationId: sendMsgEphemeralOneButtonConfirm
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              user:
                type: string
              text:
                type: string
              callbackId:
                type: string 
              textButton1:
                type: string
              textConfirmation1:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /sendMsg/ephemeral/twoButtons:
    x-swagger-router-controller: sendMsgEphemeral
    post:
      operationId: sendMsgEphemeralTwoButtons
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              user:
                type: string
              text:
                type: string
              callbackId:
                type: string 
              textButton1:
                type: string
              textButton2:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /sendMsg/ephemeral/twoButtons/confirm:
    x-swagger-router-controller: sendMsgEphemeral
    post:
      operationId: sendMsgEphemeralTwoButtonsConfirm
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              user:
                type: string
              text:
                type: string
              callbackId:
                type: string 
              textButton1:
                type: string
              textButton2:
                type: string
              textConfirmation1:
                type: string
              textConfirmation2:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
...